Feature/mod get m3g sitelogs 01a

+++ get_m3g_sitelogs.py
@@ -3,8 +3,9 @@
 
 import requests
 import subprocess
-import json
+import hashlib
 import os, glob
+import shutil
 import sys
 
 """
@@ -23,6 +24,8 @@ OPTION :
 
 * -d : delete : Delete old sitelogs in storage folder. This allows to have only
                 the last version, as version changing sitelogs changes of name.
+                
+* -m : move : Move old sitelogs into the given archive folder
 
 * -o : observatory : A four character observatory ID that will be used to filter
                      sitelogs to download. Valid values are : OVSM|OVSG|OVPF|REVOSIMA.
@@ -36,17 +39,27 @@ OPTION :
              a single observatory must be given with -o option.
              the root folder option is automatically activated (-r)
 
+* -f : force : Force download even if an identical sitelog already exists locally
+
+* -e : exclude : Site(s) you want to exclude from download. 
+                 Provide as input 4 or 9 character site codes separated with spaces
 
 EXAMPLE:
 
 ./get_m3g_sitelogs.py OUTPUTFOLDER (-d) (-o OVSM|OVSG|OVPF|REVOSIMA) (-r) (-s)
 
 2021-10-18 FL
+2023-03-21 PS
 """
 
-def get_m3g_sitelogs(sitelogsfolder, delete, observatory=None,
+def get_m3g_sitelogs(sitelogsfolder,
+                     delete,
+                     observatory=None,
                      root_folder=False,
-                     svn_mode=False):
+                     svn_mode=False,
+                     move_folder=None,
+                     force=False,
+                     exclude=[]):
     
     # the root folder option is automatically activated for the SVN mode
     if svn_mode:
@@ -56,7 +69,7 @@ def get_m3g_sitelogs(sitelogsfolder, delete, observatory=None,
     if svn_mode and not observatory:
         sys.exit("when using SVN mode (-s), a single observatory must be given with -o option")
         
-    # M3G country webservice. We use countries to filter IPGP's obs stations.
+    # M3G country webservice. We use countries to filter IPGP's obs stations.
     country_url = 'https://gnss-metadata.eu/v1/sitelog/metadata-list?country='
 
     # Observatories. First field is the counrty code in M3G's webservice.
@@ -66,7 +79,7 @@ def get_m3g_sitelogs(sitelogsfolder, delete, observatory=None,
                      'BLM' : 'OVSG',       ### St Barthelemy
                      'MAF' : 'OVSG',       ### St Martin (no station there for the moment)
                      'REU' : 'OVPF',
-                     'MTY' : 'REVOSIMA',
+                     'MYT' : 'REVOSIMA',
                      'ATF' : 'REVOSIMA'}   ### TAAF aka Terres Australes 
 
     # If an observatory ID is given, only download its sitelogs
@@ -86,9 +99,13 @@ def get_m3g_sitelogs(sitelogsfolder, delete, observatory=None,
 
     # Check that output folder exists
     if not os.path.exists(sitelogsfolder):
-        print('# ERROR : : The output folder for log file doesn\'t exist')
+        print('# ERROR : The output folder for sitelogs doesn\'t exist')
         return
 
+    if move_folder and not os.path.exists(move_folder):
+        print('# ERROR : The archive folder for old sitelogs doesn\'t exist')
+        return
+    
     # Check that obs' subfolder exists. If not, creates.
     for obs in [*observatories.values()]:
         if not root_folder:
@@ -109,6 +126,7 @@ def get_m3g_sitelogs(sitelogsfolder, delete, observatory=None,
         obs_infos = obs_infos.content.decode('utf-8')
         obs_infos = obs_infos.splitlines()
         obs_infos = obs_infos[1:] # remove header
+        obs_infos = list(reversed(obs_infos)) #list is reversed per def.
         
         if not root_folder:
             obs_path = os.path.join(sitelogsfolder, observatories[ctry])
@@ -116,15 +134,15 @@ def get_m3g_sitelogs(sitelogsfolder, delete, observatory=None,
             obs_path = sitelogsfolder
             
         # If delete, empty folders
-        if delete:
-            old_sitelogs = glob.glob(obs_path + '/*log')
-            for f in old_sitelogs:
+        if delete and not move_folder:
+            old_sitelogs_del = glob.glob(obs_path + '/*' + ctry + '*.log')
+            for f in old_sitelogs_del:
                 os.remove(f)
 
-        print("### Downloading {} ({}) sitelogs from M3G to {}".format(observatories[ctry],ctry,obs_path))
+        print("###### Downloading {} ({}) sitelogs from M3G to {}".format(observatories[ctry],ctry,obs_path))
 
         for line in obs_infos:
-            line = line.split()
+            line = line.split()            
 
             # station_name = line[0]
 
@@ -132,20 +150,52 @@ def get_m3g_sitelogs(sitelogsfolder, delete, observatory=None,
             if len(line) < 6:
                 print('### ' + line[0] + ' : not available ###')
                 continue
-
-            sitelog_url = line[5]
+            
+            sitelog_md5  = line[1]
+            sitelog_url  = line[5]
             sitelog_name = line[2]
             sitelog_local_path = os.path.join(obs_path, sitelog_name)
             Sitelog_local_paths.append(sitelog_local_path)
+            
 
-            print('### ' + sitelog_name + ' ###')
+            ### skip excluded stations
+            exclude = [e.lower() for e in exclude]
+            if sitelog_name[:4] in exclude or sitelog_name[:9] in exclude:
+                print('### ' + sitelog_name + ' skip (excluded) ###')
+                continue
+            
+            ### get the checksum for the existing sitelog, if any
+            if os.path.exists(sitelog_local_path):
+                local_file = open(sitelog_local_path,'rb')
+                local_md5 = hashlib.md5(local_file.read()).hexdigest()
+            else:
+                local_md5 = None
+            
+            if force or (local_md5 != sitelog_md5):
+                print('### ' + sitelog_name + ' download ###')
+                ##Dowload the sitelog
+                r = requests.get(sitelog_url, allow_redirects=True)
+                #print(r.status_code)
+                content = r.content #.rstrip()
+                open(sitelog_local_path, 'wb').write(content)
+                # subprocess.call(['wget',
+                #                  '--no-check-certificate',
+                #                  sitelog_url,'-q',
+                #                  '-O', sitelog_local_path])
+                
+            else:
+                print('### ' + sitelog_name + ' skip (already exists) ###')
 
-            # Dowload the sitelog
-            # r = requests.get(sitelog_url, allow_redirects=True)
-            # print(r.status_code)
-            # open(os.path.join(obs_path, sitelog_name), 'wb').write(r.content)
-            subprocess.call(['wget', '--no-check-certificate', sitelog_url, '-q', '-O', sitelog_local_path])
-    
+            ### get existing old sitelogs for moving                
+            if move_folder:
+                old_sitelogs_mv = glob.glob(obs_path + '/*' + sitelog_name[:9] + '*.log')
+                if sitelog_local_path in old_sitelogs_mv:
+                    old_sitelogs_mv.remove(sitelog_local_path)
+                for f in old_sitelogs_mv:
+                    print('### ' + os.path.basename(f) + ' moved to archive folder ###')
+                    shutil.move(f,move_folder)
+                        
+                
     if svn:
         print("### SVN add/commit of the downloaded sitelogs")
         for sitelog_local in Sitelog_local_paths:
@@ -180,13 +230,16 @@ if __name__ == '__main__':
     import argparse
 
     # Parsing Args
-    parser = argparse.ArgumentParser(description='Get all IPGP sitelogs in the last version for M3G repository')
+    parser = argparse.ArgumentParser(description='Get the last versions of the IPGP sitelogs sorted on the M3G Database')
     parser.add_argument('sitelogsfolder', type=str, help='Output folder where to store downloaded sitelogs')
-    parser.add_argument('-d', '--delete', help='Delete old sitelogs in storage folder', action='store_true')
-    parser.add_argument('-o', '--observatory', help='Download sitelog for a single observatory (OVSM|OVSG|OVPF|REVOSIMA)',
+    parser.add_argument('-d', '--delete', help='Delete old sitelogs in the output folder. This allows to have only the last version, as version changing sitelogs changes of name.', action='store_true')
+    parser.add_argument('-m', '--move', help='Move old sitelogs into the given archive folder', type=str, default=None)
+    parser.add_argument('-o', '--observatory', help='Download sitelogs for some specific observatories. Valid values are : OVSM|OVSG|OVPF|REVOSIMA',
                         type=str, choices=['OVSM', 'OVSG', 'OVPF', 'REVOSIMA'], default=None)
-    parser.add_argument('-r', '--root', help='store the sitelogs in OUTPUTFOLDER root',action='store_true',default=False)
-    parser.add_argument('-s', '--svn', help='a mode to maintain the legacy OVS SVN folder. Download the sitelog of a single obs and perform a svn commit',action='store_true',default=False)
+    parser.add_argument('-r', '--root', help='Store the sitelogs in OUTPUTFOLDER root. (per default, an observatory-specific folder is created to store the corresponding sitelogs.)',action='store_true',default=False)
+    parser.add_argument('-s', '--svn', help='A mode to maintain the legacy OVS SVN folder. Download the sitelog of a single obs and perform a svn commit. A single observatory must be given with -o option. The root folder option is automatically activated (-r)',action='store_true',default=False)
+    parser.add_argument('-f', '--force', help='Force download even if an identical sitelog already exists locally',action='store_true',default=False)
+    parser.add_argument('-e', '--exclude', help='Site(s) you want to exclude from download. Provide as input 4 or 9 character site codes separated with spaces',nargs='+',default=[])
 
     args = parser.parse_args()
     sitelogsfolder = args.sitelogsfolder
@@ -194,5 +247,9 @@ if __name__ == '__main__':
     observatory = args.observatory
     svn = args.svn
     root = args.root
+    move_folder = args.move
+    force = args.force
+    exclude = args.exclude
 
-    get_m3g_sitelogs(sitelogsfolder, delete, observatory,root,svn)
+    get_m3g_sitelogs(sitelogsfolder, delete, observatory,root,svn,
+                     move_folder, force, exclude)
